LicensePlateDetector/requirements.txt
Flask>=0.10.1
opencv-python==3.4.9.31
pillow
pytesseract
numpy
Markdown>=3.1.1